% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{github_pages_to_repo}
\alias{github_pages_to_repo}
\title{Convert GitHub Pages URLs to repository URLs}
\usage{
github_pages_to_repo(url)
}
\arguments{
\item{url}{A URL}
}
\value{
A GitHub repository URL
}
\description{
Convert GitHub Pages URLs to repository URLs
}
