% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type.R
\name{in_grammar_check_cran}
\alias{in_grammar_check_cran}
\title{Classify CRAN package as in-grammar or other}
\usage{
in_grammar_check_cran(package, cran_db = NULL)
}
\arguments{
\item{package}{Package name}

\item{cran_db}{Optional CRAN package database}
}
\value{
"in grammar" or "other"
}
\description{
Classify CRAN package as in-grammar or other
}
