% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type.R
\name{in_grammar_check_github}
\alias{in_grammar_check_github}
\title{Classify GitHub package as in-grammar or other}
\usage{
in_grammar_check_github(package, url)
}
\arguments{
\item{package}{Package name}

\item{url}{GitHub repository URL}
}
\value{
"in grammar", "other", or NA
}
\description{
Classify GitHub package as in-grammar or other
}
