% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep_type.R
\name{which_dep_cran}
\alias{which_dep_cran}
\title{Determine ggplot2 dependency for a CRAN package}
\usage{
which_dep_cran(pkg, cran_db = NULL)
}
\arguments{
\item{pkg}{Package name}

\item{cran_db}{Optional CRAN package database}
}
\value{
One of "depends", "imports", "suggests", or NA
}
\description{
Determine ggplot2 dependency for a CRAN package
}
