% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep_type.R
\name{which_dep_github}
\alias{which_dep_github}
\title{Determine ggplot2 dependency for a GitHub package}
\usage{
which_dep_github(repo_url)
}
\arguments{
\item{repo_url}{GitHub repository URL}
}
\value{
One of "depends", "imports", "suggests", or NA
}
\description{
Determine ggplot2 dependency for a GitHub package
}
