% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_components.R
\name{get_components_cran}
\alias{get_components_cran}
\title{Extract ggplot2 grammar components from a CRAN package}
\usage{
get_components_cran(pkg_name, cran_db = NULL)
}
\arguments{
\item{pkg_name}{Package name}

\item{cran_db}{Optional CRAN package database}
}
\value{
A data.frame with package, component, and function names
}
\description{
Extract ggplot2 grammar components from a CRAN package
}
