% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep_type.R
\name{which_dep_cran}
\alias{which_dep_cran}
\title{Classify ggplot2 dependency type for CRAN packages}
\usage{
which_dep_cran(pkgs, cran_db)
}
\arguments{
\item{pkgs}{Character vector of package names.}

\item{cran_db}{A data frame from tools::CRAN_package_db().}
}
\value{
Character vector with values "depends", "imports", "suggests", or NA.
}
\description{
Classify ggplot2 dependency type for CRAN packages
}
